fixed wrong expected values, added in missing test, removed whitespace and wrote white box tests for occurences.
I have fixed any of my tests that had the wrong expected value for my junit tests.

I had a test for my valid overlaps in my spreadsheet but not in my junit so I have added in a comment to mention this and i have added in the that test that was missing.

i removed most of the whitespace.

I have also wrote white box tests for the occurences part of the code.