doing tdd wrong, now I am going to fix this
I have just realized that I was not doing tdd right. I wrote the code then I wrote my tests, I believe this is wrong.

So I am deleting my code files, and I am going to write a test then the code to pass that test, this is the right way to do tdd.